+++ src/App/Witness/Database.php
@@ -8,14 +8,27 @@ use Exception;
 class Database
 {
     /**
-     * Get all the database checks.
+     * Perform checks for each database.
      *
      * @return array
      */
     public function check()
     {
-        $data['db_status'] = $this->checkDatabaseStatus();
-        $data['db_size'] = $this->checkDatabaseSize();
+        if (is_array(config('eyewitness.database_connections'))) {
+            $connections = config('eyewitness.database_connections');
+        } else {
+            $connections = [config('database.default')];
+        }
+
+        foreach ($connections as $connection) {
+            try {
+                $data[$connection]['db_status'] = $this->checkDatabaseStatus($connection);
+                $data[$connection]['db_size'] = $this->checkDatabaseSize($connection);
+            } catch (Exception $e) {
+                $data[$connection]['db_status'] = false;
+                $data[$connection]['db_size'] = -1;
+            }
+        }
 
         return $data;
     }
@@ -23,12 +36,13 @@ class Database
     /**
      * Check if the database connection is working.
      *
+     * @param  string  $connection
      * @return bool
      */
-    public function checkDatabaseStatus()
+    protected function checkDatabaseStatus($connection)
     {
         try {
-            DB::connection()->getPdo();
+            DB::connection($connection)->getPdo();
         } catch (Exception $e) {
             return false;
         }
@@ -39,17 +53,18 @@ class Database
     /**
      * Check the size of the database.
      *
+     * @param  string  $connection
      * @return bool
      */
-    public function checkDatabaseSize()
+    protected function checkDatabaseSize($connection)
     {
-        switch(DB::getDriverName()) {
+        switch(DB::connection($connection)->getDriverName()) {
             case 'mysql':
-                return $this->checkMySqlDatabaseSize();
+                return $this->checkMySqlDatabaseSize($connection);
             case 'sqlite':
-                return $this->checkSqLiteDatabaseSize();
+                return $this->checkSqLiteDatabaseSize($connection);
             case 'pgsql':
-                return $this->checkPostgresDatabaseSize();
+                return $this->checkPostgresDatabaseSize($connection);
         }
 
         return -1;
@@ -58,14 +73,16 @@ class Database
     /**
      * Return the size of the mySql database.
      *
+     * @param  string  $connection
      * @return int
-     **/
-    protected function checkMySqlDatabaseSize()
+     */
+    protected function checkMySqlDatabaseSize($connection)
     {
         try {
-            $result = (array) DB::table('information_schema.TABLES')
+            $result = (array) DB::connection($connection)
+                                ->table('information_schema.TABLES')
                                 ->select(DB::raw("sum( data_length + index_length ) AS size"))
-                                ->where('table_schema', DB::getConfig('database'))
+                                ->where('table_schema', DB::connection($connection)->getConfig('database'))
                                 ->first();
 
             return $result['size'];
@@ -77,12 +94,13 @@ class Database
     /**
      * Return the size of the Sqlite database.
      *
+     * @param  string  $connection
      * @return int
-     **/
-    protected function checkSqLiteDatabaseSize()
+     */
+    protected function checkSqLiteDatabaseSize($connection)
     {
         try {
-            $path = realpath(DB::getConfig('database'));
+            $path = realpath(DB::connection($connection)->getConfig('database'));
             if ($path !== false) {
                 return filesize($path);
             }
@@ -96,12 +114,13 @@ class Database
     /**
      * Return the size of the Postgres database.
      *
+     * @param  string  $connection
      * @return int
-     **/
-    protected function checkPostgresDatabaseSize()
+     */
+    protected function checkPostgresDatabaseSize($connection)
     {
         try {
-            $result = DB::select(DB::raw("select pg_database_size('".DB::getConfig('database')."')"));
+            $result = DB::connection($connection)->select(DB::raw("select pg_database_size('".DB::connection($connection)->getConfig('database')."')"));
             $result = (array) $result[0];
             return $result['pg_database_size'];
         } catch (Exception $e) {